Implement UseDelegate for ArgParser and CommandRunner
