+++ crates/cli/cli-components/src/impls/commands/mod.rs
@@ -1,5 +1,4 @@
 pub mod bootstrap;
 pub mod client;
-pub mod delegate;
 pub mod queries;
 pub mod start;
