+++ crates/cli/cli-components/src/impls/parse/mod.rs
@@ -1,3 +1,2 @@
-pub mod delegate;
 pub mod field;
 pub mod string;
